problem: moderators cant quickly see length of application
solution: solution: Rank each application by length. Take the rank and divide by the total number of applications to get a normalized rating. Display this in the UI somewhere (probably in the table where all the applications can be seen by moderators). Find some way to display it nicely or log an issue for a UI person to claim. #154